server command
say permissions added

+++ commands/say.js
@@ -2,8 +2,13 @@ module.exports = {
     name: 'say',
     description: "says a message",
     execute(message, args) {
-        const sayMessage = args.join(" ");
-        message.delete().catch(err => console.log(err));
-        message.channel.send(sayMessage);
+        if (message.member.hasPermission('BAN_MEMBERS')) {
+            const sayMessage = args.join(" ");
+            message.delete().catch(err => console.log(err));
+            message.channel.send(sayMessage);
+        } else {
+            message.reply('You do not have permissions for that command.')
+            return;
+        }
     }
 }
